+++ test/test.cpp
@@ -117,7 +117,7 @@ std::string serialise(const std::vector<std::string> & v)
     private:
         // When the code in loon::writer::base needs to output Loon text
         // it will call this function.
-        virtual void write(const char * utf8, int len)
+        virtual void write(const char * utf8, size_t len)
         {
             // You could write the text to a file or whatever you want.
             // Here we just append it to a string.
@@ -317,7 +317,7 @@ std::string serialise(const std::map<std::string, std::string> & m)
         std::string str;
 
     private:
-        virtual void write(const char * utf8, int len)
+        virtual void write(const char * utf8, size_t len)
         {
             str.append(utf8, len);
         }
@@ -610,7 +610,7 @@ struct variant_writer : public loon::writer::base {
     std::string str;
 
 private:
-    virtual void write(const char * utf8, int len)
+    virtual void write(const char * utf8, size_t len)
     {
         str.append(utf8, len);
     }
@@ -1122,7 +1122,7 @@ std::string serialise(const section_map & ini)
     struct ini_writer : public loon::writer::base {
         std::string str;
     private:
-        virtual void write(const char * utf8, int len)
+        virtual void write(const char * utf8, size_t len)
         {
             str.append(utf8, len);
         }
@@ -1872,7 +1872,7 @@ void test_write_loon_hex_u32()
 
     private:
         std::string str;
-        virtual void write(const char * utf8, int len)
+        virtual void write(const char * utf8, size_t len)
         {
             str.append(utf8, len);
         }
